separate server and client component on blog page
